Create Q2
Write a program that calculates how much a person would earn over a period of time

if his or her salary is one penny the first day and two pennies the second day, and continues

to double each day. The program should ask the user for the number of days.

The output earnings should be displayed in a dollar amount, not the number of pennies.

Input Validation: Do not accept a number less than 1 for the number of days worked.