Create Q4
Write a program that asks the user for a positive integer no greater than 15. The program

should then display a rectangle on the screen using the character ‘X’. The number

entered by the user will be the length of each side of the rectangle. For example, if the

user enters 5 X, the program should display the following:

XXXXX

XXXXX

XXXXX

XXXXX

XXXXX

If input is 8 Y

YYYYYYYY

YYYYYYYY

YYYYYYYY

YYYYYYYY

YYYYYYYY

YYYYYYYY

YYYYYYYY

YYYYYYYY