Create Q9
Enhance your program from Savitch_9Ed_Chap4_Prob4 by having it also print out the estimated price of the item in one and in two years from the time of the calculation. The increase in cost over one year is estimated as the inflation rate times the price at the start of the year. Define a second variable to determine the estimated cost of an item in one year,given the current price of the item and the inflation rate as arguments.